refactor(midnight-js): relocate the one useful spike assertion (#876)
The spike file had three tests: commitment-segment-independence (kept,
moved to zswap-utils.test.ts next to the function that relies on it),
fromPartsRandomized offer-population (covered by ledger-utils regression),
and ZswapOutput.new(undefined) acceptance (not exercised in production,
type system catches breakages).
